requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
magic_hour>=0.36.0
google-generativeai
moviepy>=2.0.0.dev2
edge-tts
